태그, 함수 모듈화 작업
